refactored the post component - added the post header
