--- src/elements/post-header/index.ts
@@ -1,2 +1,0 @@
-export { PostHeader } from './post-header'
-export type { PostHeaderProps } from './post-header'
