+++ src/elements/post/header/header.docs.mdx
@@ -1,9 +1,9 @@
 ---
 description: 'A PostHeader component for the top of a post.'
-labels: ['react', 'typescript', 'post-header']
+labels: ['react', 'typescript', 'header']
 ---
 
-import { PostHeader } from './post-header';
+import { PostHeader } from './header';
 
 This post-header component has two modes which toggle based on the presence of the featured image src.
