+++ src/elements/post/header/index.ts
@@ -1,0 +1,2 @@
+export { PostHeader } from './header'
+export type { PostHeaderProps } from './header'
