--- src/elements/time-ago/index.ts
@@ -1,2 +1,0 @@
-export { TimeAgo } from './time-ago'
-export type { TimeAgoProps } from './time-ago'
