--- src/elements/time-ago/time-ago.docs.mdx
@@ -1,23 +1,0 @@
----
-labels: ['react', 'typescript', 'time-ago']
-description: 'A TimeAgo component.'
----
-
-import { TimeAgo } from './time-ago';
-
-TimeAgo example:
-
-Accepts a date object and optional locale and returns a human readable time ago string.
-
-Code Snippet:
-```js
-const d = new Date()
-d.setDate(d.getDate() - 5)
-<TimeAgo datetime={d} locale='ru' />
-```
-
-
-Live Playground:
-```js live
-<TimeAgo datetime={new Date()} locale='enUS' />
-```
